requirements.txt
# Twitter API
tweepy>=4.14.0

# Discord API (using requests for simple API calls)
requests>=2.31.0

# Google Gemini AI
google-generativeai>=0.3.0

# Environment variables
python-dotenv>=1.0.0
